refactor: Nas classes entidade de Product e OrderProduct os campos referentes ao preço, que antes estavam com o tipo Double, agora são do tipo BigDecimal, visto que assim funcionaria melhor com o precision e scale. Além disso, o precision desses campos foi alterado de 6 para 10

+++ back-end/src/main/java/org/uorderflow/model/OrderProduct.java
@@ -4,6 +4,8 @@ import jakarta.persistence.*;
 import lombok.*;
 import org.uorderflow.dto.orderProduct.OrderProductCreateDTO;
 
+import java.math.BigDecimal;
+
 @Table(name = "tb_order_products")
 @Entity(name = "OrderProduct")
 @Getter
@@ -23,8 +25,8 @@ public class OrderProduct {
     @Column(nullable = false, name = "quantity")
     private Integer quantity;
 
-    @Column(nullable = false, precision = 6, scale = 2, name = "unit_price")
-    private Double unitPrice;
+    @Column(nullable = false, precision = 10, scale = 2, name = "unit_price")
+    private BigDecimal unitPrice;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(nullable = false, name = "order_id")
